+++ examples/rest-api-client-demo/src/space.ts
@@ -1,6 +1,8 @@
 import type { KintoneRestAPIClient } from "@kintone/rest-api-client";
 
 const SPACE_ID = 8;
+const GUEST_SPACE_ID = 9;
+const THREAD_ID = 8;
 
 export class Space {
   private client: KintoneRestAPIClient;
@@ -75,6 +77,99 @@ export class Space {
 
     try {
       console.log(await this.client.space.updateSpaceMembers(spaceMembers));
+    } catch (error) {
+      console.log(error);
+    }
+  }
+
+  public async updateThread() {
+    const body = "<b>This is an updated thread body</b>";
+    const name = "Updated Thread Name";
+    try {
+      console.log(
+        await this.client.space.updateThread({ id: THREAD_ID, body, name }),
+      );
+    } catch (error) {
+      console.log(error);
+    }
+  }
+
+  public async addThreadComment() {
+    const params = {
+      space: "8",
+      thread: "1",
+      comment: {
+        text: "This is a comment",
+        files: [
+          {
+            fileKey: "file1",
+            width: 100,
+          },
+        ],
+        mentions: [
+          {
+            code: "user1",
+            type: "USER" as const,
+          },
+          {
+            code: "user2",
+            type: "USER" as const,
+          },
+        ],
+      },
+    };
+    try {
+      console.log(await this.client.space.addThreadComment(params));
+    } catch (error) {
+      console.log(error);
+    }
+  }
+
+  public async addGuests() {
+    const guests = [
+      {
+        code: "guest1@example.com",
+        password: "password123",
+        timezone: "America/Los_Angeles",
+        locale: "en" as const,
+        image: "78a586f2-e73e-4a70-bec2-43976a60746e", // replace with the fileKey of the uploaded file
+        name: "John Doe",
+        company: "Company Name",
+        division: "Sales",
+        phone: "999-456-7890",
+        callto: "skypecallto",
+      },
+    ];
+    try {
+      console.log(await this.client.space.addGuests({ guests }));
+    } catch (error) {
+      console.log(error);
+    }
+  }
+
+  public async deleteGuests() {
+    const guests = ["abc1@gmail.com", "abc2@gmail.com", "abc3@gmail.com"];
+    try {
+      console.log(await this.client.space.deleteGuests({ guests }));
+    } catch (error) {
+      console.log(error);
+    }
+  }
+
+  public async updateSpaceGuests() {
+    const guests = [
+      "guestUser1@gmail.com",
+      "guestUser2@gmail.com",
+      "guestUser3@gmail.com",
+    ];
+
+    try {
+      console.log(
+        await this.client.space.updateSpaceGuests({
+          id: GUEST_SPACE_ID,
+          guests,
+        }),
+      );
     } catch (error) {
       console.log(error);
     }