+++ packages/rest-api-client/src/client/SpaceClient.ts
@@ -1,8 +1,12 @@
 import type {
   SpaceID,
+  ThreadID,
   Space,
+  ThreadComment,
   SpaceMemberForResponse,
   SpaceMemberForRequest,
+  GuestSpaceID,
+  Guest,
 } from "./types";
 import { BaseClient } from "./BaseClient";
 
@@ -43,6 +47,48 @@ export class SpaceClient extends BaseClient {
   }): Promise<{}> {
     const path = this.buildPathWithGuestSpaceId({
       endpointName: "space/members",
+    });
+    return this.client.put(path, params);
+  }
+
+  public updateThread(params: {
+    id: ThreadID;
+    name?: string;
+    body?: string;
+  }): Promise<{}> {
+    const path = this.buildPathWithGuestSpaceId({
+      endpointName: "space/thread",
+    });
+    return this.client.put(path, params);
+  }
+
+  public addThreadComment(params: ThreadComment): Promise<{ id: string }> {
+    const path = this.buildPathWithGuestSpaceId({
+      endpointName: "space/thread/comment",
+    });
+    return this.client.post(path, params);
+  }
+
+  public addGuests(params: { guests: Guest[] }): Promise<{}> {
+    const path = this.buildPathWithGuestSpaceId({
+      endpointName: "guests",
+    });
+    return this.client.post(path, params);
+  }
+
+  public deleteGuests(params: { guests: string[] }): Promise<{}> {
+    const path = this.buildPathWithGuestSpaceId({
+      endpointName: "guests",
+    });
+    return this.client.delete(path, params);
+  }
+
+  public updateSpaceGuests(params: {
+    id: GuestSpaceID;
+    guests: string[];
+  }): Promise<{}> {
+    const path = this.buildPathWithGuestSpaceId({
+      endpointName: "space/guests",
     });
     return this.client.put(path, params);
   }